+++ p2panda-rs/src/test_utils/templates/entry_templates.rs
@@ -36,13 +36,14 @@ use rstest_reuse::template;
     13,
     1,
     Some(crate::test_utils::constants::DEFAULT_HASH.parse().unwrap()),
-    Some(crate::test_utils::constants::DEFAULT_HASH.parse().unwrap()),
+    Some(crate::test_utils::fixtures::random_hash()),
     Some(crate::test_utils::fixtures::operation(
         Some(crate::test_utils::fixtures::operation_fields(crate::test_utils::constants::default_fields())),
         None,
         None
     ))
 ))]
+#[allow(unused_qualifications)]
 #[case::skiplink_can_be_omitted_when_sam_as_backlink(crate::test_utils::fixtures::entry(
     14,
     1,